+++ crates/sdk/src/client.rs
@@ -63,15 +63,6 @@ impl ProverClient {
         self.inner.execute(elf, stdin).await
     }
 
-    pub async fn prove(
-        &self,
-        pk: &SP1ProvingKey,
-        stdin: &SP1Stdin,
-    ) -> Result<SP1ProofWithPublicValues> {
-        let opts = ProofOpts::default();
-        self.inner.prove_with_options(pk, stdin, &opts).await
-    }
-
     pub fn prove<'a>(&'a self, pk: &'a SP1ProvingKey, stdin: &'a SP1Stdin) -> DynProofRequest<'a> {
         DynProofRequest::new(&*self.inner, pk, stdin, ProofOpts::default())
     }
