+++ crates/sdk/src/network-v2/mod.rs
@@ -1,6 +1,7 @@
 mod client;
 mod error;
 mod prover;
+mod retry;
 mod types;
 #[rustfmt::skip]
 mod proto;
