+++ crates/sdk/src/network-v2/prover.rs
@@ -1,28 +1,39 @@
+use std::pin::Pin;
 use std::time::{Duration, Instant};
 
 use anyhow::Result;
+use async_trait::async_trait;
 use backoff::{future::retry, Error as BackoffError, ExponentialBackoff};
 use serde::de::DeserializeOwned;
+use sp1_core_executor::{ExecutionError, ExecutionReport, SP1Context};
+use sp1_core_machine::io::SP1Stdin;
+use sp1_primitives::io::SP1PublicValues;
+use sp1_prover::{components::DefaultProverComponents, SP1Prover, SP1_CIRCUIT_VERSION};
+use sp1_prover::{SP1ProvingKey, SP1VerifyingKey};
+use sp1_stark::SP1ProverOpts;
+use std::future::{Future, IntoFuture};
+use tokio::runtime::Runtime;
+use tokio::task;
 use tokio::time::sleep;
 use tonic::Code;
 
-use sp1_core_machine::io::SP1Stdin;
-use sp1_prover::{components::DefaultProverComponents, SP1Prover, SP1_CIRCUIT_VERSION};
-use sp1_stark::SP1ProverOpts;
-
-use crate::block_on;
+use crate::mode::Mode;
+use crate::network_v2::retry::{self, with_retry};
 use crate::network_v2::{
     client::{NetworkClient, DEFAULT_PROVER_NETWORK_RPC},
     proto::network::{ExecutionStatus, FulfillmentStatus, FulfillmentStrategy, ProofMode},
     types::{HashType, RequestId, VerifyingKeyHash},
     Error,
 };
+use crate::opts::ProofOpts;
+use crate::proof::SP1ProofWithPublicValues;
+use crate::prover::Prover;
+use crate::provers::SP1VerificationError;
+use crate::request::{DEFAULT_CYCLE_LIMIT, DEFAULT_TIMEOUT};
+use crate::{block_on, verify};
 
 /// The default fulfillment strategy to use for proof requests.
 pub const DEFAULT_FULFILLMENT_STRATEGY: FulfillmentStrategy = FulfillmentStrategy::Hosted;
-
-/// The default timeout for a proof request to be fulfilled (4 hours).
-pub const DEFAULT_TIMEOUT_SECS: u64 = 14400;
 
 /// The minimum allowed timeout for a proof request to be fulfilled (10 seconds).
 pub const MIN_TIMEOUT_SECS: u64 = 10;
@@ -30,362 +41,426 @@ pub const MIN_TIMEOUT_SECS: u64 = 10;
 /// The maximum allowed timeout for a proof request to be fulfilled (24 hours).
 pub const MAX_TIMEOUT_SECS: u64 = 86400;
 
-/// The default cycle limit for a proof request if simulation and the cycle limit is not explicitly
-/// set.
-pub const DEFAULT_CYCLE_LIMIT: u64 = 100_000_000;
-
 /// The number of seconds to wait between checking the status of a proof request.
-pub const STATUS_INTERVAL_SECS: u64 = 2;
+pub const STATUS_CHECK_INTERVAL_SECS: u64 = 2;
 
-// /// An implementation of [crate::ProverClient] that can generate proofs on a remote RPC server.
-// pub struct NetworkProver {
-//     client: NetworkClient,
-//     local_prover: CpuProver,
-// }
+/// An implementation of [crate::ProverClient] that can generate proofs on a remote RPC server.
+pub struct NetworkProver {
+    prover: SP1Prover<DefaultProverComponents>,
+    network_client: NetworkClient,
+}
 
-// impl NetworkProver {
-//     /// Creates a new `NetworkProver` with the given private key.
-//     pub fn new(private_key: &str) -> Self {
-//         let version = SP1_CIRCUIT_VERSION;
-//         log::info!("Client circuit version: {}", version);
+pub struct NetworkProverBuilder {
+    rpc_url: Option<String>,
+    private_key: Option<String>,
+}
 
-//         Self { client: NetworkClient::new(private_key), local_prover: CpuProver::new() }
-//     }
+impl NetworkProver {
+    /// Creates a new `NetworkProver` with the given private key.
+    pub fn new(rpc_url: String, private_key: String) -> Self {
+        Self {
+            prover: SP1Prover::new(),
+            network_client: NetworkClient::new(&private_key).with_rpc_url(rpc_url),
+        }
+    }
 
-//     /// Sets up the proving key and verifying key for the given ELF.
-//     pub fn setup(&self, elf: &[u8]) -> (SP1ProvingKey, SP1VerifyingKey) {
-//         self.local_prover.setup(elf)
-//     }
+    /// Sets up the proving key and verifying key for the given ELF.
+    pub fn with_mode(mut self, mode: Mode) -> Self {
+        self.network_client.mode = mode.into();
+        self
+    }
 
-//     /// Sets the RPC URL for the prover network.
-//     ///
-//     /// This configures the endpoint that will be used for all network operations.
-//     /// If not set, the default RPC URL will be used.
-//     pub fn with_rpc_url(mut self, rpc_url: impl Into<String>) -> Self {
-//         self.client = self.client.with_rpc_url(rpc_url);
-//         self
-//     }
+    /// Sets the RPC URL for the prover network.
+    ///
+    /// This configures the endpoint that will be used for all network operations.
+    /// If not set, the default RPC URL will be used.
+    pub fn timeout_secs(mut self, secs: u64) -> Self {
+        self.network_client.timeout_secs = Some(secs);
+        self
+    }
 
-//     /// Creates a new network prover builder. See [`NetworkProverBuilder`] for more details.
-//     pub fn builder() -> NetworkProverBuilder {
-//         NetworkProverBuilder::default()
-//     }
+    /// Sets the cycle limit for the prover network.
+    ///
+    /// See `get_cycle_limit` for more details the final cycle limit is determined.
+    pub fn cycle_limit(mut self, limit: u64) -> Self {
+        self.network_client.cycle_limit = Some(limit);
+        self
+    }
 
-//     /// Gets the clamped timeout in seconds from the provided options.
-//     fn get_timeout_secs(timeout: Option<Duration>) -> u64 {
-//         timeout
-//             .map(|d| d.as_secs().clamp(MIN_TIMEOUT_SECS, MAX_TIMEOUT_SECS))
-//             .unwrap_or(DEFAULT_TIMEOUT_SECS)
-//     }
+    /// Skips simulation when determining the cycle limit.
+    ///
+    /// See `get_cycle_limit` for more details the final cycle limit is determined.
+    pub fn skip_simulation(mut self, skip: bool) -> Self {
+        self.network_client.skip_simulation = skip;
+        self
+    }
 
-//     /// Get the cycle limit to used for a proof request.
-//     ///
-//     /// The cycle limit is determined according to the following priority:
-//     /// 1. If a cycle limit was explicitly set, use the specified value
-//     /// 2. If simulation is enabled (default), calculate the limit by simulating
-//     /// 3. Otherwise use the default cycle limit
-//     #[allow(clippy::must_use_candidate)]
-//     fn get_cycle_limit(
-//         &self,
-//         elf: &[u8],
-//         stdin: &SP1Stdin,
-//         cycle_limit: Option<u64>,
-//         skip_simulation: bool,
-//     ) -> Result<u64, Error> {
-//         // If cycle_limit was explicitly set, use it.
-//         if let Some(limit) = cycle_limit {
-//             return Ok(limit);
-//         }
+    /// Sets the fulfillment strategy for the prover network.
+    ///
+    /// See `request_proof` for more details the final cycle limit is determined.
+    pub fn strategy(mut self, strategy: FulfillmentStrategy) -> Self {
+        self.network_client.fulfillment_strategy = Some(strategy);
+        self
+    }
 
-//         // If simulation is enabled (default), simulate to get the limit.
-//         if !skip_simulation {
-//             let (_, report) = self
-//                 .local_prover
-//                 .sp1_prover()
-//                 .execute(elf, stdin, Default::default())
-//                 .map_err(|_| Error::SimulationFailed)?;
-//             Ok(report.total_instruction_count())
-//         } else {
-//             // Skip simulation was set but no explicit cycle limit, use default.
-//             Ok(DEFAULT_CYCLE_LIMIT)
-//         }
-//     }
+    /// Get the cycle limit to used for a proof request.
+    ///
+    /// The cycle limit is determined according to the following priority:
+    /// 1. If a cycle limit was explicitly set, use the specified value
+    /// 2. If simulation is enabled (default), calculate the limit by simulating
+    /// 3. Otherwise use the default cycle limit
+    #[allow(clippy::must_use_candidate)]
+    fn get_cycle_limit(
+        &self,
+        elf: &[u8],
+        stdin: &SP1Stdin,
+        cycle_limit: Option<u64>,
+        skip_simulation: bool,
+    ) -> Result<u64, Error> {
+        // If cycle_limit was explicitly set, use it.
+        if let Some(limit) = cycle_limit {
+            return Ok(limit);
+        }
 
-//     /// Registers a program if it is not already registered.
-//     pub async fn register_program(
-//         &self,
-//         vk: &SP1VerifyingKey,
-//         elf: &[u8],
-//     ) -> Result<VerifyingKeyHash> {
-//         self.client.register_program(vk, elf).await
-//     }
+        // If simulation is enabled (default), simulate to get the limit.
+        if !skip_simulation {
+            let (_, report) = self
+                .prover
+                .execute(elf, stdin, Default::default())
+                .map_err(|_| Error::SimulationFailed)?;
+            Ok(report.total_instruction_count())
+        } else {
+            // Skip simulation was set but no explicit cycle limit, use default.
+            Ok(DEFAULT_CYCLE_LIMIT)
+        }
+    }
 
-//     /// Requests a proof from the prover network, returning the request ID.
-//     #[allow(clippy::too_many_arguments)]
-//     pub async fn request_proof(
-//         &self,
-//         vk_hash: &VerifyingKeyHash,
-//         stdin: &SP1Stdin,
-//         version: &str,
-//         mode: ProofMode,
-//         strategy: FulfillmentStrategy,
-//         timeout_secs: u64,
-//         cycle_limit: u64,
-//     ) -> Result<RequestId> {
-//         // Request the proof with retries.
-//         let (tx_hash, request_id) = with_retry(
-//             || async {
-//                 self.client
-//                     .request_proof(
-//                         vk_hash,
-//                         stdin,
-//                         version,
-//                         mode,
-//                         strategy,
-//                         timeout_secs,
-//                         cycle_limit,
-//                     )
-//                     .await
-//             },
-//             timeout_secs,
-//             "requesting proof",
-//         )
-//         .await?;
+    /// Registers a program if it is not already registered.
+    pub async fn register_program(
+        &self,
+        vk: &SP1VerifyingKey,
+        elf: &[u8],
+    ) -> Result<VerifyingKeyHash> {
+        self.network_client.register_program(vk, elf).await
+    }
 
-//         log::info!("Created request {} in transaction {}", request_id, tx_hash);
+    /// Requests a proof from the prover network, returning the request ID.
+    #[allow(clippy::too_many_arguments)]
+    pub async fn request_proof(
+        &self,
+        vk_hash: &VerifyingKeyHash,
+        stdin: &SP1Stdin,
+        version: &str,
+        mode: Mode,
+        strategy: FulfillmentStrategy,
+        timeout_secs: u64,
+        cycle_limit: u64,
+    ) -> Result<RequestId> {
+        // Request the proof with retries.
+        let (tx_hash, request_id) = retry::with_retry(
+            || async {
+                self.network_client
+                    .request_proof(
+                        vk_hash,
+                        stdin,
+                        version,
+                        mode.into(),
+                        strategy,
+                        timeout_secs,
+                        cycle_limit,
+                    )
+                    .await
+            },
+            timeout_secs,
+            "requesting proof",
+        )
+        .await?;
 
-//         if self.client.rpc_url() == DEFAULT_PROVER_NETWORK_RPC {
-//             log::info!("View in explorer: {}", request_id.explorer_url());
-//         }
+        log::info!("Created request {} in transaction {}", request_id, tx_hash);
 
-//         Ok(request_id)
-//     }
+        if self.network_client.rpc_url() == DEFAULT_PROVER_NETWORK_RPC {
+            log::info!("View in explorer: {}", request_id.explorer_url());
+        }
 
-//     /// Waits for the proof request to be fulfilled by the prover network.
-//     ///
-//     /// The proof request must have already been submitted. This function will return a
-//     /// `RequestTimedOut` error if the request does not received a response within the timeout.
-//     pub async fn wait_proof<P: DeserializeOwned>(
-//         &self,
-//         request_id: &RequestId,
-//         timeout_secs: u64,
-//     ) -> Result<P, Error> {
-//         let mut is_assigned = false;
-//         let start_time = Instant::now();
-//         let timeout = Duration::from_secs(timeout_secs);
+        Ok(request_id)
+    }
 
-//         loop {
-//             // Check if we've exceeded the timeout.
-//             if start_time.elapsed() > timeout {
-//                 return Err(Error::RequestTimedOut { request_id: request_id.clone() });
-//             }
-//             let remaining_timeout = timeout.saturating_sub(start_time.elapsed());
+    /// Waits for the proof request to be fulfilled by the prover network.
+    ///
+    /// The proof request must have already been submitted. This function will return a
+    /// `RequestTimedOut` error if the request does not received a response within the timeout.
+    pub async fn wait_proof<P: DeserializeOwned>(
+        &self,
+        request_id: &RequestId,
+        timeout_secs: u64,
+    ) -> Result<P, Error> {
+        let mut is_assigned = false;
+        let start_time = Instant::now();
+        let timeout = Duration::from_secs(timeout_secs);
 
-//             // Get the status with retries.
-//             let (status, maybe_proof) = with_retry(
-//                 || async { self.client.get_proof_request_status(request_id).await },
-//                 remaining_timeout.as_secs(),
-//                 "getting proof status",
-//             )
-//             .await?;
+        loop {
+            // Check if we've exceeded the timeout.
+            if start_time.elapsed() > timeout {
+                return Err(Error::RequestTimedOut { request_id: request_id.clone() });
+            }
+            let remaining_timeout = timeout.saturating_sub(start_time.elapsed());
 
-//             // Check the deadline.
-//             if status.deadline < Instant::now().elapsed().as_secs() {
-//                 return Err(Error::RequestTimedOut { request_id: request_id.clone() });
-//             }
+            // Get the status with retries.
+            let (status, maybe_proof) = with_retry(
+                || async { self.network_client.get_proof_request_status(request_id).await },
+                remaining_timeout.as_secs(),
+                "getting proof status",
+            )
+            .await?;
 
-//             // Check the execution status.
-//             if let Ok(ExecutionStatus::Unexecutable) =
-//                 ExecutionStatus::try_from(status.execution_status)
-//             {
-//                 return Err(Error::RequestUnexecutable { request_id: request_id.clone() });
-//             }
+            // Check the deadline.
+            if status.deadline < Instant::now().elapsed().as_secs() {
+                return Err(Error::RequestTimedOut { request_id: request_id.clone() });
+            }
 
-//             // Check the fulfillment status.
-//             match FulfillmentStatus::try_from(status.fulfillment_status) {
-//                 Ok(FulfillmentStatus::Fulfilled) => {
-//                     return Ok(maybe_proof.unwrap());
-//                 }
-//                 Ok(FulfillmentStatus::Assigned) => {
-//                     if !is_assigned {
-//                         log::info!("Proof request assigned, proving...");
-//                         is_assigned = true;
-//                     }
-//                 }
-//                 Ok(FulfillmentStatus::Unfulfillable) => {
-//                     return Err(Error::RequestUnfulfillable { request_id: request_id.clone() });
-//                 }
-//                 _ => {}
-//             }
+            // Check the execution status.
+            if let Ok(ExecutionStatus::Unexecutable) =
+                ExecutionStatus::try_from(status.execution_status)
+            {
+                return Err(Error::RequestUnexecutable { request_id: request_id.clone() });
+            }
 
-//             sleep(Duration::from_secs(STATUS_INTERVAL_SECS)).await;
-//         }
-//     }
+            // Check the fulfillment status.
+            match FulfillmentStatus::try_from(status.fulfillment_status) {
+                Ok(FulfillmentStatus::Fulfilled) => {
+                    return Ok(maybe_proof.unwrap());
+                }
+                Ok(FulfillmentStatus::Assigned) => {
+                    if !is_assigned {
+                        log::info!("Proof request assigned, proving...");
+                        is_assigned = true;
+                    }
+                }
+                Ok(FulfillmentStatus::Unfulfillable) => {
+                    return Err(Error::RequestUnfulfillable { request_id: request_id.clone() });
+                }
+                _ => {}
+            }
 
-//     /// Requests a proof from the prover network and waits for it to be returned.
-//     pub async fn prove(
-//         &self,
-//         pk: &SP1ProvingKey,
-//         stdin: SP1Stdin,
-//         opts: ProofOpts,
-//         kind: SP1ProofKind,
-//     ) -> Result<SP1ProofWithPublicValues, Error> {
-//         // Ensure the program is registered.
-//         let vk_hash = self.register_program(&pk.vk, &pk.elf).await?;
+            sleep(Duration::from_secs(STATUS_CHECK_INTERVAL_SECS)).await;
+        }
+    }
 
-//         // Get the configured settings.
-//         let version = SP1_CIRCUIT_VERSION;
-//         let mode = kind.into();
-//         let strategy = opts.fulfillment_strategy.unwrap_or(DEFAULT_FULFILLMENT_STRATEGY);
-//         let timeout_secs = Self::get_timeout_secs(opts.timeout);
-//         let cycle_limit =
-//             self.get_cycle_limit(&pk.elf, &stdin, opts.cycle_limit, opts.skip_simulation)?;
+    pub fn prove_with_options<'a>(
+        &'a self,
+        pk: &'a SP1ProvingKey,
+        stdin: &'a SP1Stdin,
+    ) -> NetworkProofRequest<'a> {
+        NetworkProofRequest::new(self, pk, stdin)
+    }
 
-//         // Request the proof.
-//         let request_id = self
-//             .request_proof(&vk_hash, &stdin, version, mode, strategy, timeout_secs, cycle_limit)
-//             .await?;
+    /// Creates a new network prover builder. See [`NetworkProverBuilder`] for more details.
+    pub fn builder() -> NetworkProverBuilder {
+        NetworkProverBuilder::new()
+    }
+}
 
-//         // Wait for the proof to be generated.
-//         self.wait_proof(&request_id, timeout_secs).await
-//     }
-// }
+impl NetworkProverBuilder {
+    /// Creates a new network prover builder.
+    pub fn new() -> Self {
+        Self { rpc_url: None, private_key: None }
+    }
 
-// impl Prover<DefaultProverComponents> for NetworkProver {
-//     fn id(&self) -> ProverType {
-//         ProverType::Network
-//     }
+    /// Sets the RPC URL for the prover network.
+    ///
+    /// This configures the endpoint that will be used for all network operations.
+    /// If not set, the default RPC URL will be used.
+    pub fn rpc_url(mut self, url: String) -> Self {
+        self.rpc_url = Some(url);
+        self
+    }
 
-//     fn setup(&self, elf: &[u8]) -> (SP1ProvingKey, SP1VerifyingKey) {
-//         self.local_prover.setup(elf)
-//     }
+    /// Sets the private key to use for the prover network.
+    ///
+    /// This is required and must be set before building the prover.
+    pub fn private_key(mut self, key: String) -> Self {
+        self.private_key = Some(key);
+        self
+    }
 
-//     fn sp1_prover(&self) -> &SP1Prover {
-//         self.local_prover.sp1_prover()
-//     }
+    /// Builds the prover with the given configuration.
+    pub fn build(self) -> NetworkProver {
+        NetworkProver::new(
+            self.rpc_url.unwrap_or_else(|| DEFAULT_PROVER_NETWORK_RPC.to_string()),
+            self.private_key.expect("private key is required"),
+        )
+    }
+}
 
-//     fn prove<'a>(
-//         &'a self,
-//         pk: &SP1ProvingKey,
-//         stdin: SP1Stdin,
-//         opts: ProofOpts,
-//         context: SP1Context<'a>,
-//         kind: SP1ProofKind,
-//     ) -> Result<SP1ProofWithPublicValues> {
-//         warn_if_not_default(&opts.sp1_prover_opts, &context);
-//         block_on(self.prove(pk, stdin, opts, kind)).map_err(Into::into)
-//     }
-// }
+pub struct NetworkProofRequest<'a> {
+    prover: &'a NetworkProver,
+    pk: &'a SP1ProvingKey,
+    stdin: &'a SP1Stdin,
+    version: String,
+    mode: Mode,
+    fulfillment_strategy: FulfillmentStrategy,
+    timeout: u64,
+    cycle_limit: Option<u64>,
+    skip_simulation: bool,
+}
 
-// /// Warns if `opts` or `context` are not default values, since they are currently unsupported.
-// fn warn_if_not_default(opts: &SP1ProverOpts, context: &SP1Context) {
-//     let _guard = tracing::warn_span!("network_prover").entered();
-//     if opts != &SP1ProverOpts::default() {
-//         tracing::warn!("non-default opts will be ignored: {:?}", opts.core_opts);
-//         tracing::warn!("custom SP1ProverOpts are currently unsupported by the network prover");
-//     }
-//     // Exhaustive match is done to ensure we update the warnings if the types change.
-//     let SP1Context { hook_registry, subproof_verifier, .. } = context;
-//     if hook_registry.is_some() {
-//         tracing::warn!("non-default context.hook_registry will be ignored: {:?}", hook_registry);
-//         tracing::warn!("custom runtime hooks are currently unsupported by the network prover");
-//         tracing::warn!("proving may fail due to missing hooks");
-//     }
-//     if subproof_verifier.is_some() {
-//         tracing::warn!("non-default context.subproof_verifier will be ignored");
-//         tracing::warn!("custom subproof verifiers are currently unsupported by the network prover");
-//     }
-// }
+impl<'a> NetworkProofRequest<'a> {
+    pub fn new(prover: &'a NetworkProver, pk: &'a SP1ProvingKey, stdin: &'a SP1Stdin) -> Self {
+        Self {
+            prover,
+            pk,
+            stdin,
+            version: SP1_CIRCUIT_VERSION.to_owned(),
+            mode: Mode::default(),
+            fulfillment_strategy: DEFAULT_FULFILLMENT_STRATEGY,
+            timeout: DEFAULT_TIMEOUT,
+            cycle_limit: None,
+            skip_simulation: false,
+        }
+    }
 
-// impl From<SP1ProofKind> for ProofMode {
-//     fn from(value: SP1ProofKind) -> Self {
-//         match value {
-//             SP1ProofKind::Core => Self::Core,
-//             SP1ProofKind::Compressed => Self::Compressed,
-//             SP1ProofKind::Plonk => Self::Plonk,
-//             SP1ProofKind::Groth16 => Self::Groth16,
-//         }
-//     }
-// }
+    pub fn with_mode(mut self, mode: Mode) -> Self {
+        self.mode = mode;
+        self
+    }
 
-// /// Execute an async operation with exponential backoff retries.
-// pub async fn with_retry<T, F, Fut>(
-//     operation: F,
-//     timeout_secs: u64,
-//     operation_name: &str,
-// ) -> Result<T>
-// where
-//     F: Fn() -> Fut,
-//     Fut: std::future::Future<Output = Result<T>>,
-// {
-//     let timeout = Duration::from_secs(timeout_secs);
-//     let backoff = ExponentialBackoff {
-//         initial_interval: Duration::from_secs(1),
-//         max_interval: Duration::from_secs(120),
-//         max_elapsed_time: Some(timeout),
-//         ..Default::default()
-//     };
+    pub fn with_timeout(mut self, timeout: u64) -> Self {
+        self.timeout = Some(timeout);
+        self
+    }
 
-//     retry(backoff, || async {
-//         match operation().await {
-//             Ok(result) => Ok(result),
-//             Err(e) => {
-//                 // Check for tonic status errors.
-//                 if let Some(status) = e.downcast_ref::<tonic::Status>() {
-//                     match status.code() {
-//                         Code::Unavailable => {
-//                             log::warn!(
-//                                 "Network temporarily unavailable when {} due to {}, retrying...",
-//                                 operation_name,
-//                                 status.message(),
-//                             );
-//                             Err(BackoffError::transient(e))
-//                         }
-//                         Code::NotFound => {
-//                             log::error!(
-//                                 "{} not found due to {}",
-//                                 operation_name,
-//                                 status.message(),
-//                             );
-//                             Err(BackoffError::permanent(e))
-//                         }
-//                         _ => {
-//                             log::error!(
-//                                 "Permanent error encountered when {}: {} ({})",
-//                                 operation_name,
-//                                 status.message(),
-//                                 status.code()
-//                             );
-//                             Err(BackoffError::permanent(e))
-//                         }
-//                     }
-//                 } else {
-//                     // Check for common transport errors.
-//                     let error_msg = e.to_string().to_lowercase();
-//                     let is_transient = error_msg.contains("tls handshake") ||
-//                         error_msg.contains("dns error") ||
-//                         error_msg.contains("connection reset") ||
-//                         error_msg.contains("broken pipe") ||
-//                         error_msg.contains("transport error") ||
-//                         error_msg.contains("failed to lookup");
+    pub fn with_strategy(mut self, strategy: FulfillmentStrategy) -> Self {
+        self.fulfillment_strategy = Some(strategy);
+        self
+    }
 
-//                     if is_transient {
-//                         log::warn!(
-//                             "Transient transport error when {}: {}, retrying...",
-//                             operation_name,
-//                             error_msg
-//                         );
-//                         Err(BackoffError::transient(e))
-//                     } else {
-//                         log::error!("Permanent error when {}: {}", operation_name, error_msg);
-//                         Err(BackoffError::permanent(e))
-//                     }
-//                 }
-//             }
-//         }
-//     })
-//     .await
-// }
+    pub fn with_cycle_limit(mut self, cycle_limit: u64) -> Self {
+        self.cycle_limit = Some(cycle_limit);
+        self
+    }
+
+    async fn run(self) -> Result<SP1ProofWithPublicValues> {
+        // Ensure the program is registered.
+        let vk_hash = self.prover.register_program(&self.pk.vk, &self.pk.elf).await?;
+
+        // Get the cycle limit.
+        let cycle_limit = self.prover.get_cycle_limit(
+            &self.pk.elf,
+            &self.stdin,
+            self.cycle_limit,
+            self.skip_simulation,
+        )?;
+
+        // Request the proof.
+        let request_id = self
+            .prover
+            .request_proof(
+                &vk_hash,
+                &self.stdin,
+                &self.version,
+                self.mode,
+                self.fulfillment_strategy,
+                self.timeout,
+                cycle_limit,
+            )
+            .await?;
+
+        // Wait for proof generation.
+        let proof: SP1ProofWithPublicValues = self
+            .prover
+            .wait_proof(&request_id, self.timeout)
+            .await
+            .map_err(|e| anyhow::anyhow!("Failed to wait for proof: {}", e))?;
+
+        Ok(proof)
+    }
+}
+
+impl<'a> IntoFuture for NetworkProofRequest<'a> {
+    type Output = Result<SP1ProofWithPublicValues>;
+    type IntoFuture = Pin<Box<dyn Future<Output = Self::Output> + Send + 'a>>;
+
+    fn into_future(self) -> Self::IntoFuture {
+        Box::pin(async move { self.run().await })
+    }
+}
+
+#[async_trait]
+impl Prover for NetworkProver {
+    async fn setup(&self, elf: &[u8]) -> (SP1ProvingKey, SP1VerifyingKey) {
+        task::spawn_blocking(move || self.prover.setup(elf)).await.unwrap()
+    }
+
+    #[cfg(feature = "blocking")]
+    fn setup_sync(&self, elf: &[u8]) -> (SP1ProvingKey, SP1VerifyingKey) {
+        self.prover.setup(elf)
+    }
+
+    async fn execute(
+        &self,
+        elf: &[u8],
+        stdin: &SP1Stdin,
+    ) -> Result<(SP1PublicValues, ExecutionReport), ExecutionError> {
+        task::spawn_blocking(move || self.prover.execute(elf, stdin, SP1Context::default()))
+            .await
+            .unwrap()
+    }
+
+    #[cfg(feature = "blocking")]
+    fn execute_sync(
+        &self,
+        elf: &[u8],
+        stdin: &SP1Stdin,
+    ) -> Result<(SP1PublicValues, ExecutionReport), ExecutionError> {
+        self.prover.execute(elf, stdin, SP1Context::default())
+    }
+
+    async fn prove_with_options(
+        &self,
+        pk: &SP1ProvingKey,
+        stdin: &SP1Stdin,
+        opts: &ProofOpts,
+    ) -> Result<SP1ProofWithPublicValues> {
+        let request = NetworkProofRequest::new(self, pk, stdin)
+            .with_mode(opts.mode)
+            .with_timeout(opts.timeout)
+            .with_cycle_limit(opts.cycle_limit);
+        request.run().await
+    }
+
+    #[cfg(feature = "blocking")]
+    fn prove_with_options_sync(
+        &self,
+        pk: &SP1ProvingKey,
+        stdin: &SP1Stdin,
+        opts: &ProofOpts,
+    ) -> Result<SP1ProofWithPublicValues> {
+        let request = NetworkProofRequest::new(self, pk, stdin)
+            .with_mode(opts.mode)
+            .with_timeout(opts.timeout)
+            .with_cycle_limit(opts.cycle_limit);
+        Runtime::new().unwrap().block_on(request.run())
+    }
+
+    async fn verify(
+        &self,
+        proof: &SP1ProofWithPublicValues,
+        vk: &SP1VerifyingKey,
+    ) -> Result<(), SP1VerificationError> {
+        task::spawn_blocking(move || verify::verify(&self.prover, SP1_CIRCUIT_VERSION, proof, vk))
+            .await
+            .unwrap()
+    }
+
+    #[cfg(feature = "blocking")]
+    fn verify_sync(
+        &self,
+        proof: &SP1ProofWithPublicValues,
+        vk: &SP1VerifyingKey,
+    ) -> Result<(), SP1VerificationError> {
+        verify::verify(&self.prover, SP1_CIRCUIT_VERSION, proof, vk)
+    }
+}
 
 // #[cfg(test)]
 // mod tests {
